fix credit button delayed enable
