+++ Assets/Core/UI/MainMenu/Scripts/StartButton.cs
@@ -9,7 +9,6 @@ using UnityEngine.UI;
 public class StartButton : MonoBehaviour
 {
     [SerializeField] private float _waitDuration;
-    [SerializeField] private float _enableIn;
     private Button _button;
 
     public delegate void OnStartButton();
@@ -25,14 +24,7 @@ public class StartButton : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        _button.enabled = false;
-        StartCoroutine(EnableButton());
-    }
 
-    // Update is called once per frame
-    void Update()
-    {
-        
     }
 
     public void ButtonPress() {
@@ -43,10 +35,5 @@ public class StartButton : MonoBehaviour
         onStartTransition?.Invoke();
         yield return new WaitForSeconds(_waitDuration);
         onStartButton?.Invoke();
-    }
-
-    IEnumerator EnableButton() {
-        yield return new WaitForSeconds(_enableIn);
-        _button.enabled = true;
     }
 }
