+++ Assets/Core/UI/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 381158ec50bf8b044b3e7d480ccf7603
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
