+++ Assets/Core/UI/Scripts/WaitToEnable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d98a1256548bddf4eb34ddf4ee5b1cc9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
